backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.26
botocore==1.38.26
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
jmespath==1.0.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
s3transfer==0.13.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wheel==0.45.1
